#31 - Added generator for UI mounting of Specterr Web + Fixed namespace issues with Rails generators that we have.
