--- lib/generators/specterr/specterr_generator.rb
@@ -1,17 +1,0 @@
-require 'rails/generators/base'
-
-class SpecterrGenerator < Rails::Generators::Base
-  source_root File.expand_path('../templates', __FILE__)
-
-  def create_specterr_file
-    create_file "config/specterr.yml", default_postgresql_config
-  end
-
-  def default_postgresql_config
-    "adapter: postgresql\ndatabase: <database_name>\nusername: <user_name>\npassword: <password>\nhost: <hostname>\nport: <port>"
-  end
-
-  def default_sqlite_config
-    "adapter: sqlite3,\npool:  5,\ntimeout: 5000,\ndatabase: db/specterr.db"
-  end
-end
